pyproject.toml
[tool.poetry]
name = "ipydocs"
version = "0.0.0"
description = "Interactive Python Documentation"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8"
numpydoc = "^1.5.0"
toml = "^0.10.2"
click = "^8.1.6"
Jinja2 = "^3.1.2"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.278"
build = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ipydocs = "ipydocs.__main__:commands"
